+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### X.X.X (In progress)
 **SDK**
-- **Feature:** Added `sendMessageToContact` interface in `ChatMessageBridgeDispatcher` for sending message to a single contact. HostApp can get the message using `MessageToContact` object.
+- **Feature:** Added `sendMessageToContact` interface in `ChatBridgeDispatcher` for sending message to a single contact. HostApp can get the message using `MessageToContact` object.
 
 **Sample App**
 - **Feature:** Added first-time launching screen before downloading any miniapp.